+++ src/Constants.php
@@ -14,7 +14,7 @@ namespace Awesomite\StackTrace;
 /**
  * @internal
  */
-class Constants
+final class Constants
 {
     const KEY_FILE_OBJECT    = '__awesomite_file_object';
     const KEY_ARGS_CONVERTED = '__awesomite_args_converted';